requirements.txt
# Core
numpy
pandas
scipy
scikit-learn
networkx
pyarrow
fastparquet

# ML & IR
sentence-transformers
faiss-cpu
mlflow
joblib

# Dashboard & Viz
streamlit
streamlit-pyvis
pyvis
matplotlib
seaborn

# LLM
google-generativeai
python-dotenv
hydra-core
pyyaml